requirements.txt
flask-sqlalchemy
flask-migrate
pymysql
python-dotenv
flask-pymongo
jsonify
flask-login
flask-bcrypt